Moving InlineIconText over.
