+++ app/src/main/java/com/adammcneilly/pocketleague/ui/GameDetailContent.kt
@@ -16,8 +16,8 @@ import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
+import com.adammcneilly.pocketleague.android.designsystem.components.InlineIconText
 import com.adammcneilly.pocketleague.core.displaymodels.GameDetailDisplayModel
-import com.adammcneilly.pocketleague.ui.composables.components.InlineIconText
 import com.adammcneilly.pocketleague.ui.composables.stats.PlayerStatsTable
 
 /**
