+++ app/src/main/java/com/adammcneilly/pocketleague/ui/composables/match/MatchListItem.kt
@@ -12,9 +12,9 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
+import com.adammcneilly.pocketleague.android.designsystem.components.InlineIconText
 import com.adammcneilly.pocketleague.core.displaymodels.MatchDetailDisplayModel
 import com.adammcneilly.pocketleague.core.displaymodels.MatchTeamResultDisplayModel
-import com.adammcneilly.pocketleague.ui.composables.components.InlineIconText
 
 /**
  * Renders a [displayModel] inside of an individual list item along side other matches
